Revert "refactor(containermanager): move shared types and constants to providers common file"
This reverts commit 59faeb71f863b62c1629ed7b257fa910107329a7.
